pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "autosvc"
version = "0.1.0"
description = "Local automotive diagnostic service with a client-agnostic core, CLI/TUI frontends, and optional daemon."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-can>=4.3.1",
  "textual>=0.52.1",
]

[project.scripts]
autosvc = "autosvc.apps.cli:main"
autosvc-ecu-sim = "autosvc.emulator.ecu_sim:main"

[tool.hatch.build.targets.wheel]
packages = ["autosvc"]
